backend-py/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
langchain>=0.1.5
langchain-openai>=0.0.6
anthropic>=0.3.10
